requirements.txt
playwright
beautifulsoup4
python-whois
lightgbm
pandas
joblib